Got rid of "directive" package since it seems like it's not going to be useful. Moved its only class (DirectiveType) to the top-level package with the other "type" enums.
