+++ src/main/java/com/gnopai/ji65/scanner/Scanner.java
@@ -1,7 +1,7 @@
 package com.gnopai.ji65.scanner;
 
+import com.gnopai.ji65.DirectiveType;
 import com.gnopai.ji65.InstructionType;
-import com.gnopai.ji65.directive.DirectiveType;
 
 import java.util.List;
