+++ src/test/java/com/gnopai/ji65/assembler/AssemblerTest.java
@@ -1,9 +1,9 @@
 package com.gnopai.ji65.assembler;
 
+import com.gnopai.ji65.DirectiveType;
 import com.gnopai.ji65.InstructionType;
 import com.gnopai.ji65.config.ProgramConfig;
 import com.gnopai.ji65.config.SegmentConfig;
-import com.gnopai.ji65.directive.DirectiveType;
 import com.gnopai.ji65.parser.statement.DirectiveStatement;
 import com.gnopai.ji65.parser.statement.InstructionStatement;
 import com.gnopai.ji65.parser.statement.LabelStatement;
